# Reduce CPU overhead of FusionDefinitionWrapper