This PR lowers the CPU overhead of every nvFuser fusion execution in Thunder.

This is done by improving the runtime performance of two things inside `FusionDefinitionWrapper.__call__`:
1. `to_descriptors` performance is improved
2. NVTX annotation uses a different function relying on the `nvtx` package for annotation with less Python overhead

Here's a breakdown of improvements to `to_descriptors` for a microbenchmark calling nvFuser fusion for a different number of Tensor inputs:
1. Original: 83.871 µs
2. Move "if Tensor" to top: 68.574 µs
3. Use Tensor instead of torch.Tensor: 66.805 µs
4. Inline get_tensor_descriptor: 64.152 µs
5. Use TensorProxy._shape instead of .shape: 29.003 µs

`thunder.core.profile.add_markers` is slow (constant ~30 µs overhead) because it uses `torch.profiler.record_function`, a Torch Autograd specific annotation. We're executing nvFuser fusions in a no_grad context so the record_function doesn't do anything useful for us here. `thunder.core.profile.annotate_for_profile` is a better alternative because it simply calls fast `nvtx.annotate`.

Here are graphs before and after this change (we want to see more of the yellow part):
(before)
![image](https://github.com/user-attachments/assets/34180697-6b18-4502-ad52-013d8acd26f2)
"fusion_wrapper_time_other" includes nvtx annotations and some other attribute accesses and writes we do in the `__call__` method.
(after)
![image](https://github.com/user-attachments/assets/66a10576-1201-4767-9def-622bda277323)

And here's a table of speedups for a different number of input Tensors:
| Number of input Tensors | Total time FusionWrapper | to_descriptors | NVTX Annotation + usability attribute writes |
|-------------------------|--------------------------|----------------|-------------------------------------------------------------------|
| 1                       |                    2.86x |          2.08x |                                                             5.91x |
| 2                       |                    2.50x |          2.38x |                                                             4.88x |
| 4                       |                    2.35x |          2.55x |                                                             4.65x |
| 8                       |                    2.23x |          2.67x |                                                             5.04x |
| 16                      |                    1.96x |          2.75x |                                                             4.04x |
| 32                      |                    1.74x |          2.80x |                                                             3.99x |
| 64                      |                    1.61x |          2.83x |                                                             4.00x |

Script to collect the numbers: https://gist.github.com/IvanYashchuk/1276fd7781f18c8003cbc688e8cc64df

cc @tfogal